==> 2_scripts/0D_Calculo de máximos por ciclo para Aysén.R <==
library(dplyr)
library(tidyverse)

sites <-
  readr::read_rds("joined_data.rds") |>
  janitor::clean_names()
sites <-
  sites |>
  dplyr::mutate(
    adultos_totales = prom_ho + prom_am,
    prop.juv = prom_juv/(prom_juv + adultos_totales)) |>
  dplyr::select(codigo, especie, nombre_region,  fecha_primer_muestreo, fecha_declaracion,
                biomasa_centro, peso_promedio, numero_total_peces,
                prom_juv, prom_ho, prom_am, adultos_totales,
                prop.juv, temperatura_promedio,
                salinidad_promedio) |>
  dplyr::mutate(
    ddiff =  if_else(
      is.na(lag(fecha_declaracion)),
      interval(fecha_declaracion-days(7), ymd(fecha_declaracion)),
      interval(ymd(lag(fecha_declaracion)), ymd(fecha_declaracion))
    ),
    ddiff = ddiff/weeks(1),
    .by = c(codigo, especie)) |>
  dplyr::mutate(new_cycle = if_else(ddiff > 7 | fecha_declaracion == ymd(min(fecha_declaracion)), 1, 0), .by = c(codigo, especie)) |>
  dplyr::mutate(cycle = cumsum(new_cycle), .by = c(codigo, especie)) |>
  dplyr::mutate(
    month = month(fecha_declaracion),
    year = year(fecha_declaracion),
    carga_total = (adultos_totales + prom_juv)*40,
    estacion = case_when(
      month == 1|month ==2 |month == 12~ "summer",
      month == 3|month ==4 |month == 5~ "autumn",
      month == 6|month ==7 |month == 8~ "winter",
      month == 9|month ==10 |month == 11~ "spring",
      
    )
  )




meanMax.Cycles <-
  sites |>
  dplyr::filter(
    especie %in% c("SALMON DEL ATLANTICO", "TRUCHA ARCOIRIS")) |>
  dplyr::filter(cycle > 1 & cycle < max(cycle),
                .by = c(codigo, especie)) |>
  summarize(
    max.adultos_totales = max(adultos_totales, na.rm = T),
    max.carga_total = max(carga_total, na.rm = T),
    .by = c(codigo, nombre_region,
            especie,cycle)) |>
  filter(max.carga_total >1) |>
  left_join(
    sites[,c(1:2,6:15,16:22)],
    by = c("codigo","especie",
           "cycle",
           "max.carga_total" = "carga_total"))


max.summary <-
  meanMax.Cycles |>
  summarize(
    max.adultos_totales = mean(max.adultos_totales, na.rm = T),
    max.carga_total = mean(max.carga_total, na.rm = T),
    log_max.adultos_totales = mean(log1p(max.adultos_totales), na.rm = T),
    log_max.carga_total = mean(log1p(max.carga_total), na.rm = T),
    peso_promedio = mean(peso_promedio, na.rm = T),
    numero_total_peces = mean(numero_total_peces, na.rm = T),
    prom_juv = mean(prom_juv, na.rm = T),
    prop.juv = mean(prop.juv, na.rm = T),
    temperatura_promedio = mean(temperatura_promedio, na.rm = T),
    salinidad_promedio = mean(salinidad_promedio, na.rm = T),
    .by=c(codigo, nombre_region,especie, cycle)) |>
  summarize(
    max.carga_total = mean(max.carga_total, na.rm = T),
    max.adultos_totales = mean(max.adultos_totales, na.rm = T),
    log_max.carga_total = mean(log_max.carga_total, na.rm = T),
    log_max.adultos_totales = mean(log_max.adultos_totales, na.rm = T),
    peso_promedio = mean(peso_promedio, na.rm = T),
    numero_total_peces = mean(numero_total_peces, na.rm = T),
    prom_juv = mean(prom_juv, na.rm = T),
    prop.juv = mean(prop.juv, na.rm = T),
    temperatura_promedio = mean(temperatura_promedio, na.rm = T),
    salinidad_promedio = mean(salinidad_promedio, na.rm = T),
    .by=c(codigo,nombre_region, especie)) |>
  left_join(full_shape[,c(3,22)],
            by = c("codigo" = "N_CODIGOCE")) |> 
  filter(nombre_region == "XI REGION") |>
  arrange(codigo)
